refactor: refactor study group services into StudyGroupService, StudyGroupAdminService
