+++ src/main/java/com/samsamhajo/deepground/studyGroup/controller/StudyGroupController.java
@@ -7,18 +7,13 @@ import com.samsamhajo.deepground.member.repository.MemberRepository;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupCommentRequest;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupCommentResponse;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupInviteRequest;
-import com.samsamhajo.deepground.studyGroup.dto.StudyGroupKickRequest;
-import com.samsamhajo.deepground.studyGroup.dto.StudyGroupMemberSummary;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupParticipationResponse;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupReplyRequest;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupReplyResponse;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupSearchRequest;
 import com.samsamhajo.deepground.studyGroup.dto.StudyGroupUpdateRequest;
-import com.samsamhajo.deepground.studyGroup.service.StudyGroupAcceptService;
 import com.samsamhajo.deepground.studyGroup.service.StudyGroupCommentService;
-import com.samsamhajo.deepground.studyGroup.service.StudyGroupDeleteService;
 import com.samsamhajo.deepground.studyGroup.service.StudyGroupInviteService;
-import com.samsamhajo.deepground.studyGroup.service.StudyGroupReplyService;
 import com.samsamhajo.deepground.studyGroup.service.StudyGroupService;
 import com.samsamhajo.deepground.studyGroup.success.StudyGroupSuccessCode;
 import com.samsamhajo.deepground.member.entity.Member;
@@ -37,14 +32,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/study-group")
 public class StudyGroupController {
-
   private final StudyGroupService studyGroupService;
-  private final StudyGroupDeleteService deleteService;
-  private final StudyGroupInviteService inviteService;
-  private final MemberRepository memberRepository;
-  private final StudyGroupReplyService replyService;
-  private final StudyGroupCommentService commentService;
-  private final StudyGroupAcceptService acceptService;
 
   @GetMapping("/{studyGroupId}")
   public ResponseEntity<SuccessResponse<?>> getStudyGroupDetail(
@@ -88,26 +76,6 @@ public class StudyGroupController {
         .body(SuccessResponse.of(StudyGroupSuccessCode.CREATE_SUCCESS, response));
   }
 
-
-
-
-  @PostMapping("/{studyGroupId}/accept/{targetMemberId}")
-  public ResponseEntity<SuccessResponse<?>> acceptMember(
-      @PathVariable Long studyGroupId,
-      @PathVariable Long targetMemberId,
-      @AuthenticationPrincipal CustomUserDetails customUserDetails
-  ) {
-    Member requester = customUserDetails.getMember();
-
-    acceptService.acceptMember(studyGroupId, targetMemberId, requester);
-
-    return ResponseEntity
-        .status(StudyGroupSuccessCode.UPDATE_SUCCESS.getStatus())
-        .body(SuccessResponse.of(StudyGroupSuccessCode.UPDATE_SUCCESS));
-  }
-
-
-
   @GetMapping("/joined")
   public ResponseEntity<SuccessResponse<List<StudyGroupParticipationResponse>>> getMyParticipatedGroups(
       @AuthenticationPrincipal CustomUserDetails customUserDetails
@@ -146,56 +114,11 @@ public class StudyGroupController {
     Member requester = customUserDetails.getMember();
     GlobalLogger.info("스터디 소프트 삭제 요청", studyGroupId, requester.getEmail());
 
-    deleteService.softDeleteStudyGroup(studyGroupId, requester);
+    studyGroupService.softDeleteStudyGroup(studyGroupId, requester);
 
     return ResponseEntity
         .status(StudyGroupSuccessCode.DELETE_SUCCESS.getStatus())
         .body(SuccessResponse.of(StudyGroupSuccessCode.DELETE_SUCCESS));
-  }
-
-  @PostMapping("/invite")
-  public ResponseEntity<SuccessResponse<?>> inviteByEmail(
-      @RequestBody @Valid StudyGroupInviteRequest request,
-      @AuthenticationPrincipal CustomUserDetails customUserDetails
-  ) {
-    Member invitor = customUserDetails.getMember();
-
-    inviteService.inviteByEmail(invitor, request);
-    return ResponseEntity
-        .status(StudyGroupSuccessCode.CREATE_SUCCESS.getStatus())
-        .body(SuccessResponse.of(StudyGroupSuccessCode.CREATE_SUCCESS));
-  }
-
-
-
-
-
-  @PostMapping("/comments")
-  public ResponseEntity<SuccessResponse<StudyGroupCommentResponse>> writeComment(
-      @RequestBody @Valid StudyGroupCommentRequest request,
-      @AuthenticationPrincipal CustomUserDetails customUserDetails
-  ) {
-    Member member = customUserDetails.getMember();
-    GlobalLogger.info("스터디 댓글 작성 요청", member.getEmail(), request.getStudyGroupId());
-
-    StudyGroupCommentResponse response = commentService.writeComment(request, member.getId());
-
-    return ResponseEntity
-        .status(StudyGroupSuccessCode.COMMENT_CREATE_SUCCESS.getStatus())
-        .body(SuccessResponse.of(StudyGroupSuccessCode.COMMENT_CREATE_SUCCESS, response));
-  }
-
-  @PostMapping("/replies")
-  public ResponseEntity<SuccessResponse<StudyGroupReplyResponse>> writeReply(
-      @AuthenticationPrincipal CustomUserDetails customUserDetails,
-      @RequestBody @Valid StudyGroupReplyRequest request
-  ) {
-    Member member = customUserDetails.getMember();
-
-    StudyGroupReplyResponse response = replyService.writeReply(member, request);
-    return ResponseEntity
-        .status(StudyGroupSuccessCode.CREATE_SUCCESS.getStatus())
-        .body(SuccessResponse.of(StudyGroupSuccessCode.CREATE_SUCCESS, response));
   }
 
     @PatchMapping("/{studyGroupId}")
