+++ src/main/java/com/samsamhajo/deepground/studyGroup/controller/StudyGroupInviteController.java
@@ -1,0 +1,38 @@
+package com.samsamhajo.deepground.studyGroup.controller;
+
+import com.samsamhajo.deepground.auth.security.CustomUserDetails;
+import com.samsamhajo.deepground.global.success.SuccessResponse;
+import com.samsamhajo.deepground.member.entity.Member;
+import com.samsamhajo.deepground.studyGroup.dto.StudyGroupInviteRequest;
+import com.samsamhajo.deepground.studyGroup.service.StudyGroupInviteService;
+import com.samsamhajo.deepground.studyGroup.success.StudyGroupSuccessCode;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/study-group")
+public class StudyGroupInviteController {
+
+  private final StudyGroupInviteService inviteService;
+
+  @PostMapping("/invite")
+  public ResponseEntity<SuccessResponse<?>> inviteByEmail(
+      @RequestBody @Valid StudyGroupInviteRequest request,
+      @AuthenticationPrincipal CustomUserDetails customUserDetails
+  ) {
+    Member invitor = customUserDetails.getMember();
+
+    inviteService.inviteByEmail(invitor, request);
+    return ResponseEntity
+        .status(StudyGroupSuccessCode.CREATE_SUCCESS.getStatus())
+        .body(SuccessResponse.of(StudyGroupSuccessCode.CREATE_SUCCESS));
+  }
+
+}
